+++ src/main/java/function/FunctionInterpreter.java
@@ -4,7 +4,7 @@ package function;
 import java.util.ArrayList;
 
 public class FunctionInterpreter {
-    public static float evaluate(double[] inputs, ArrayList<Float> numbers, ArrayList<ArrayList<Integer>> function){
+    public static double evaluate(double[] inputs, ArrayList<Double> numbers, ArrayList<ArrayList<Integer>> function){
 
 
         return 0;
