+++ src/main/java/function/Zahl.java
@@ -3,8 +3,8 @@ package function;
 public interface Zahl {
     public boolean isFinal();
     public boolean isComplex();
-    public float getValue();
-    public void setValue(float v);
+    public double getValue();
+    public void setValue(double v);
     public Zahl multiply(Zahl zahl);
     public Zahl add(Zahl zahl);
 }
